--- internal/cmd/start_interactive_test.go
@@ -1,92 +1,0 @@
-package cmd
-
-import (
-	"os"
-	"testing"
-
-	"github.com/stretchr/testify/require"
-	"golang.org/x/term"
-)
-
-func TestCmdStart_InteractiveMode(t *testing.T) {
-	t.Run("Should return error when no DAG specified and not in TTY", func(t *testing.T) {
-		// This test verifies that when no DAG is specified and we're not in an
-		// interactive terminal, the command returns an appropriate error
-		// Since tests don't run in a TTY, this should fail with "DAG file path is required"
-		t.Skip("This test is for documentation purposes - actual testing requires TTY simulation")
-	})
-
-	t.Run("Should work with DAG path provided", func(t *testing.T) {
-		// Create a test DAG file
-		tmpDir := t.TempDir()
-		dagFile := tmpDir + "/test.yaml"
-		dagContent := `
-name: test-dag
-steps:
-  - name: step1
-    command: echo test
-`
-		err := os.WriteFile(dagFile, []byte(dagContent), 0644)
-		require.NoError(t, err)
-
-		// Test that providing a DAG path works as before
-		cmd := CmdStart()
-		cmd.SetArgs([]string{dagFile})
-		
-		// The command might succeed or fail depending on the test environment
-		// What we're testing is that it accepts the DAG file argument
-		_ = cmd.Execute()
-		// If there's an error, it should not be about missing DAG file
-		// (The actual execution might fail for other reasons in test environment)
-	})
-
-	t.Run("Terminal detection works correctly", func(_ *testing.T) {
-		// This test just verifies that the term.IsTerminal function is available
-		// and can be called without panicking
-		_ = term.IsTerminal(int(os.Stdin.Fd()))
-	})
-}
-
-func TestCmdStart_BackwardCompatibility(t *testing.T) {
-	t.Run("Should accept parameters after --", func(t *testing.T) {
-		tmpDir := t.TempDir()
-		dagFile := tmpDir + "/test-params.yaml"
-		dagContent := `
-name: test-params
-params: KEY1=default1 KEY2=default2
-steps:
-  - name: step1
-    command: echo $KEY1 $KEY2
-`
-		err := os.WriteFile(dagFile, []byte(dagContent), 0644)
-		require.NoError(t, err)
-
-		cmd := CmdStart()
-		cmd.SetArgs([]string{dagFile, "--", "KEY1=value1", "KEY2=value2"})
-		
-		// Execute will fail due to missing context setup, but we're testing
-		// that the command accepts the arguments
-		_ = cmd.Execute()
-	})
-
-	t.Run("Should accept --params flag", func(t *testing.T) {
-		tmpDir := t.TempDir()
-		dagFile := tmpDir + "/test-params-flag.yaml"
-		dagContent := `
-name: test-params-flag
-params: KEY=default
-steps:
-  - name: step1
-    command: echo $KEY
-`
-		err := os.WriteFile(dagFile, []byte(dagContent), 0644)
-		require.NoError(t, err)
-
-		cmd := CmdStart()
-		cmd.SetArgs([]string{dagFile, "--params", "KEY=value"})
-		
-		// Execute will fail due to missing context setup, but we're testing
-		// that the command accepts the arguments
-		_ = cmd.Execute()
-	})
-}
